requirements.txt
discord.py[voice]
psutil
motor
pymongo
aiohttp
asyncpraw
thefuzz[speedup]
git+https://github.com/PythonistaGuild/Wavelink.git
python-youtube
lyricsgenius